test(admin): split the backup tests so they run in parallel
`backup.test.ts` had grown to 65 tests and 18.8s of test time — more than the rest of the package's backup tests put together, and all of it serialized in one file because vitest parallelizes across files rather than within them. It is now three files along the lines the tests already fell into:

    create_backup.test.ts           9 tests, 3.8s   what a finished backup looks like
    create_backup_failures.test.ts 32 tests, 7.4s   what happens when something goes wrong
    validate_backup.test.ts        24 tests, 6.8s   reading a backup back

Same 65 tests, unchanged apart from two renames that follow the setup into `test/backup.ts`: `machineConfig` is now `MACHINE_CONFIG` and `logger()` is `mockLogger()`. Running the package's backup tests goes from 20.9s to 10.8s of wall clock.

The shared fixtures move to `test/backup.ts`, which already held the workspace builders: the machine config, the two logger helpers, the backup-directory name, `createValidBackup` for tests that need a backup rather than a run that makes one, and `mockRoomToWorkIn` for the free-space and flush mocks. That last one asserts that the calling file mocked the two packages it reaches into, because `vi.mock` is per-file and cannot move into a helper — without the check a forgotten mock would fail somewhere far from the cause.

The longest file is now `validate_backup.test.ts` at 6.8s, nearly all of it building a full workspace and signing a real backup for each of 24 tests. A fixture that writes a backup directory directly, without a database or an election package, would cut that further; left alone here because it changes what those tests are standing on.
