+++ apps/admin/backend/test/backup.ts
@@ -5,13 +5,24 @@ import {
   makeTemporaryDirectory,
   readElectionGeneralDefinition,
 } from '@votingworks/fixtures';
-import { mockBaseLogger } from '@votingworks/logging';
+import {
+  LogEventId,
+  mockBaseLogger,
+  MockBaseLogger,
+} from '@votingworks/logging';
 import { zipFile } from '@votingworks/test-utils';
 import {
   DEFAULT_SYSTEM_SETTINGS,
   ElectionPackageFileName,
   LATEST_METADATA,
 } from '@votingworks/types';
+import { assert } from '@votingworks/basics';
+import { getDiskSpaceSummary } from '@votingworks/backend';
+import { syncFilesystem } from '@votingworks/usb-drive';
+import { generateElectionBasedSubfolderName } from '@votingworks/utils';
+import { BackupStep, createBackup } from '../src/backup/create_backup.js';
+import { BackupManifest } from '../src/backup/manifest.js';
+import { MachineConfig } from '../src/types.js';
 import { createWorkspace, Workspace } from '../src/util/workspace.js';
 
 /**
@@ -76,4 +87,82 @@ export async function makeConfiguredWorkspace(): Promise<Workspace> {
   writeFileSync(join(workspace.path, 'machine_mode'), 'host');
 
   return workspace;
+}
+
+/**
+ * The machine a backup fixture claims to have been made on.
+ */
+export const MACHINE_CONFIG: MachineConfig = {
+  machineId: 'AD-1234',
+  codeVersion: '1.2.3',
+};
+
+/**
+ * A logger whose calls a test can inspect.
+ */
+export function mockLogger(): MockBaseLogger {
+  return mockBaseLogger({ fn: vi.fn });
+}
+
+/**
+ * The stages a run reported, in order, from what it logged.
+ */
+export function loggedSteps(log: MockBaseLogger): BackupStep[] {
+  return vi
+    .mocked(log.log)
+    .mock.calls.filter(
+      ([eventId]) => eventId === LogEventId.BackupCreateProgress
+    )
+    .map(([, , logData]) => (logData as { step: BackupStep }).step);
+}
+
+/**
+ * The directory name a backup of the fixture election takes on a drive.
+ */
+export function expectedBackupDirectoryName(): string {
+  const electionDefinition = readElectionGeneralDefinition();
+  return generateElectionBasedSubfolderName(
+    electionDefinition.election,
+    electionDefinition.ballotHash
+  );
+}
+
+/**
+ * Puts a backup on a drive for tests that are about reading one back rather
+ * than writing one.
+ */
+export async function createValidBackup(): Promise<{
+  backupDirectoryPath: string;
+  manifest: BackupManifest;
+}> {
+  const workspace = await makeConfiguredWorkspace();
+  const result = await createBackup({
+    workspace,
+    targetDirectoryPath: makeTemporaryDirectory(),
+    machineConfig: MACHINE_CONFIG,
+    logger: mockLogger(),
+  });
+  return result.unsafeUnwrap();
+}
+
+/**
+ * Gives every disk a terabyte free and makes flushing succeed, so a test only
+ * has to say what it wants to go wrong.
+ *
+ * The caller must have mocked `@votingworks/backend` and
+ * `@votingworks/usb-drive` — `vi.mock` is per-file and cannot be moved here.
+ */
+export function mockRoomToWorkIn(): void {
+  assert(
+    vi.isMockFunction(getDiskSpaceSummary) && vi.isMockFunction(syncFilesystem),
+    "mockRoomToWorkIn needs the calling test file to vi.mock '@votingworks/backend' " +
+      "and '@votingworks/usb-drive'"
+  );
+  // 1 TB free everywhere, in the 1K blocks `df` reports.
+  vi.mocked(getDiskSpaceSummary).mockResolvedValue({
+    total: 1_000_000_000,
+    used: 0,
+    available: 1_000_000_000,
+  });
+  vi.mocked(syncFilesystem).mockResolvedValue();
 }
